requirements.txt
Django==4.2.4
djangorestframework==3.14.0
requests==2.31.0
markdown==3.4.4
django-filter==23.2
beautifulsoup4==4.12.2
djangorestframework-simplejwt==5.3.0
pika==1.3.2
beautifulsoup4==4.12.2
lxml==4.9.3